genpolicy: allow specifying layer cache file
Add --layers-cache-file-path flag to allow the user to
specify where the cache file for the container layers
is saved. This allows e.g. to have one cache file
independent of the user's working directory.

NOTE: This patch file has been heavily modified so that it
applies to the microsoft/kata-containers fork.
Still, of the original commit is adopted at some point,
this patch can be dropped.

Signed-off-by: Leonard Cohnen <lc@edgeless.systems>

Signed-off-by: Paul Meyer <49727155+katexochen@users.noreply.github.com>